Added progress display (Warning!)
Addead a message to display progress but it slows down the execution of the script. Must find a other way to display progress that dont render every iteration.

I dont want to launch other process to create a loading animation because if it crashes it wont kill the created process